pyproject.toml
[tool.poetry]
name = "fiber"
version = "1.0.0"
description = ""
authors = ["Nikita Chuprin <nikita.chuprin@hardwario.com>"]
packages = [{include = "fiber"}]

[tool.poetry.scripts]
fiber = "fiber:run"
test = 'test:main'

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.2"
paho-mqtt = "^1.6.1"
click = "^8.1.7"
schedule = "^1.2.1"
pyyaml = "^6.0.1"
netifaces = "^0.11.0"
smbus2 = "^0.4.3"
pyserial = "^3.5"
pydantic = "^2.6.4"
gpiod = "^2.1.3"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core>=1.8.1", "wheel"]
build-backend = "poetry.core.masonry.api"
